nix/packages/config-normalizer/pyproject.toml
[project]
name = "config-normalizer"
version = "0.1.0"
description = "MCP server that converts config files (JSON/YAML/TOML/INI) to LLM-friendly XML"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0",
    "tomli>=2.0",
]

[project.scripts]
config-normalizer = "config_normalizer:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
